extract notice builders and constents from frontend to a shred lib
