+++ src/channels/terminal_ui/mod.rs
@@ -10,7 +10,6 @@ mod execution_browser;
 mod history_cells;
 mod markdown;
 pub mod panes;
-pub(crate) mod protocol;
 mod run;
 mod syntect_highlight;
 mod text_format;
