+++ src/execution/execution_jobs.rs
@@ -17,12 +17,12 @@ use tokio::sync::{mpsc, RwLock};
 use tokio::task::{AbortHandle, JoinHandle};
 
 use crate::bus::{BusMessage, InboundMessage, TelemetryEvent, METADATA_SYNTHETIC_JOB_FOLLOWUP};
-use crate::channels::terminal::{
-    build_execution_job_notice, build_execution_job_started_notice, build_execution_stream_notice,
-};
 use crate::execution::error::ExecutionError;
 use crate::execution::{persist_successful_execution_run, PersistSuccessfulExecutionRunParams};
 use crate::execution::{CwdPolicy, ExecutionHarness, RunEvent, RunResult, RunSpec, SessionId};
+use crate::protocol::{
+    build_execution_job_notice, build_execution_job_started_notice, build_execution_stream_notice,
+};
 
 const JOB_RUNNING: u8 = 1;
 const JOB_COMPLETED: u8 = 2;
